requirements.txt
confluent-kafka==2.5.3
dnspython==2.6.1
pymongo==4.9.1
python-dotenv==1.0.1
